// src/app/security/operator-status/sys-security/models/rpdma93.model.ts
import { MapData } from './map-data.model';

/**
 * Model class Rpdma93
 * Path: screenbean/operstatsyssecur
 * Model: com::uhc::aarp::fox::domain::screenbean::operstatsyssecur::Rpdma93
 */
export class Rpdma93 {
  m93cmnd = '';
  m93mem1 = '';
  m93mem2 = '';
  m93mem3 = '';
  m93cno1 = '';
  m93cno2 = '';
  m93cno3 = '';
  m93cno4 = '';
  m93cno5 = '';
  m93cno6 = '';
  m93com1 = '';
  m93com2 = '';
  m93pdat = '';
  m93site = '';
  m93div = '';
  m93loc = '';
  m93oi01 = '';
  m93na01 = '';
  m93cl01 = '';
  m93on01 = '';
  m93of01 = '';
  m93et01 = '';
  m93oi02 = '';
  m93na02 = '';
  m93cl02 = '';
  m93on02 = '';
  m93of02 = '';
  m93et02 = '';
  m93oi03 = '';
  m93na03 = '';
  m93cl03 = '';
  m93on03 = '';
  m93of03 = '';
  m93et03 = '';
  m93oi04 = '';
  m93na04 = '';
  m93cl04 = '';
  m93on04 = '';
  m93of04 = '';
  m93et04 = '';
  m93oi05 = '';
  m93na05 = '';
  m93cl05 = '';
  m93on05 = '';
  m93of05 = '';
  m93et05 = '';
  m93oi06 = '';
  m93na06 = '';
  m93cl06 = '';
  m93on06 = '';
  m93of06 = '';
  m93et06 = '';
  m93oi07 = '';
  m93na07 = '';
  m93cl07 = '';
  m93on07 = '';
  m93of07 = '';
  m93et07 = '';
  m93oi08 = '';
  m93na08 = '';
  m93cl08 = '';
  m93on08 = '';
  m93of08 = '';
  m93et08 = '';
  m93oi09 = '';
  m93na09 = '';
  m93cl09 = '';
  m93on09 = '';
  m93of09 = '';
  m93et09 = '';
  m93oi10 = '';
  m93na10 = '';
  m93cl10 = '';
  m93on10 = '';
  m93of10 = '';
  m93et10 = '';
  m93oi11 = '';
  m93na11 = '';
  m93cl11 = '';
  m93on11 = '';
  m93of11 = '';
  m93et11 = '';
  m93oi12 = '';
  m93na12 = '';
  m93cl12 = '';
  m93on12 = '';
  m93of12 = '';
  m93et12 = '';
  m93oi13 = '';
  m93na13 = '';
  m93cl13 = '';
  m93on13 = '';
  m93of13 = '';
  m93et13 = '';
  m93oi14 = '';
  m93na14 = '';
  m93cl14 = '';
  m93on14 = '';
  m93of14 = '';
  m93et14 = '';
  m93err1 = '';
  m93pf3 = 'PF3 - NEXT OPERATOR';
  m93pf4 = 'PF4 - PREV OPERATORS';
  mapDatas: MapData[] = [];
}